# swagger/admin/swagger.yaml
definitions:
  dto.ContactInfo:
    properties:
      bin:
        description: БИН организации
        example: "5189011425"
        type: string
      email:
        description: Имейл
        example: oleg@tinkoff.ru
        type: string
      first_name:
        description: Имя
        example: Oleg
        type: string
      last_name:
        description: Фамилия
        example: Tinkoff
        type: string
      location:
        description: Адрес организации
        example: Bldg. 26, 38A, 2 Khutorskaya str., Moscow, Russia.
        type: string
      phone:
        description: Телефон
        example: "87777777777"
        type: string
      web_address:
        description: Веб адрес организации
        example: https://www.tinkoff.ru
        type: string
    type: object
...
  dto.OrderRequest:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      amount:
        type: string
      channel:
        type: string
      cityId:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      goods:
        items:
          $ref: '#/definitions/models.Goods'
        type: array
      isDelivery:
        type: boolean
      loanLength:
        type: integer
      paymentMethod:
        type: string
      paymentPartners:
        items:
          $ref: '#/definitions/models.PaymentPartners'
        type: array
      productType:
        type: string
      redirectUrl:
        type: string
      systemCode:
        type: string
      verificationId:
        type: string
      verificationSmsCode:
        type: string
      verificationSmsDateTime:
        type: string
    type: object
  dto.Orders:
    properties:
      orders:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      total:
        type: integer
    type: object
  dto.PartnerRequest:
    properties:
      code:
        description: Код партнера
        example: airba_pay
        type: string
      commission:
        description: Коммиссия
        example: 5
        type: integer
      company_name:
        description: Название компании
        example: Airba Pay
        type: string
      contact:
        $ref: '#/definitions/dto.ContactInfo'
        description: Контактные данные
      enabled:
        description: Включен
        example: true
        type: boolean
      logo_url:
        description: Веб адрес компании
        example: https://emotionsgroup.kz/uploads/7f47072c77ad20f58acb9e7114cfeb5c.jpg
        type: string
      password:
        description: Пароль
        example: Password123
        type: string
      url:
        $ref: '#/definitions/dto.URL'
        description: Адреса на обновления АПИ
      username:
        description: Имя пользователя
        example: SuperRinat
        type: string
    type: object
  dto.Partners:
    properties:
      partners:
        items:
          $ref: '#/definitions/models.Partner'
        type: array
      total:
        type: integer
    type: object
  dto.URL:
    properties:
      create:
        description: Урл который создает заказ на стороне партнера
        example: https://
        type: string
      update:
        description: Урл который обновляет заказ на стороне партнера
        example: https://
        type: string
    type: object
  httperrors.Details:
    properties:
      code:
        description: application-определенный код ошибки
        type: integer
      message:
        description: application-level сообщение, для дебага
        type: string
      status:
        description: сообщение пользовательского уровня
        type: string
    type: object
  httperrors.Response:
    properties:
      error:
        $ref: '#/definitions/httperrors.Details'
      validation:
        additionalProperties:
          type: string
        description: ошибки валидации
        type: object
    type: object
  models.Address:
    properties:
      delivery:
        type: string
      pickupPoint:
        type: string
    type: object
  models.Contact:
    properties:
      email:
        type: string
      mobileNumber:
        type: string
    type: object
  models.ContactInfo:
    properties:
      bin:
        description: БИН организации
        example: "5189011425"
        type: string
      email:
        description: Имейл
        example: oleg@tinkoff.ru
        type: string
      first_name:
        description: Имя
        example: Oleg
        type: string
      last_name:
        description: Фамилия
        example: Tinkoff
        type: string
      location:
        description: Адрес организации
        example: Bldg. 26, 38A, 2 Khutorskaya str., Moscow, Russia.
        type: string
      phone:
        description: Телефон
        example: "87777777777"
        type: string
      web_address:
        description: Веб адрес организации
        example: https://www.tinkoff.ru
        type: string
    type: object
  models.Customer:
    properties:
      contact:
        $ref: '#/definitions/models.Contact'
      firstName:
        type: string
      lastName:
        type: string
      middleName:
        type: string
      taxCode:
        type: string
    type: object
  models.Goods:
    properties:
      brand:
        type: string
      category:
        type: string
      image:
        type: string
      model:
        type: string
      price:
        type: string
    type: object
...
  models.Order:
    properties:
      address:
        $ref: '#/definitions/models.Address'
      channel:
        type: string
      created_at:
        type: string
      customer:
        $ref: '#/definitions/models.Customer'
      goods:
        items:
          $ref: '#/definitions/models.Goods'
        type: array
      id:
        type: string
      isDelivery:
        type: boolean
      loanLength:
        type: string
      orderState:
        type: string
      paymentMethod:
        type: string
      paymentPartners:
        items:
          $ref: '#/definitions/models.PaymentPartners'
        type: array
      productType:
        type: string
      redirectUrl:
        type: string
      referenceId:
        type: string
      salesPlace:
        type: string
      system_code:
        type: string
      totalCost:
        type: string
      updated_at:
        type: string
      verificationId:
        type: string
      verificationSmsCode:
        type: string
      verificationSmsDateTime:
        type: string
    type: object
  models.Partner:
    properties:
      code:
        type: string
      commission:
        type: integer
      company_name:
        type: string
      contact:
        $ref: '#/definitions/models.ContactInfo'
      created_at:
        type: string
      enabled:
        type: boolean
      id:
        type: string
      logo_url:
        type: string
      password:
        type: string
      updated_at:
        type: string
      url:
        $ref: '#/definitions/models.URL'
      username:
        type: string
    type: object
  models.PaymentPartners:
    properties:
      code:
        type: string
    type: object
  models.Response:
    properties:
      id:
        type: string
      status:
        type: string
    type: object
  models.URL:
    properties:
      create:
        type: string
      update:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admins/orders:
    get:
      description: Получение заказов
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination limit
        in: query
        name: limit
        type: integer
      - description: pagination skip
        in: query
        name: skip
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Orders'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение заказов
      tags:
      - Orders
  /admins/orders/{id}:
    get:
      consumes:
      - application/json
      description: Получение заказа
      parameters:
      - description: id of the order
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.OrderRequest'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение заказа
      tags:
      - Orders
    put:
      consumes:
      - application/json
      description: Обновление заказа
      parameters:
      - description: body
        in: body
        name: market
        required: true
        schema:
          $ref: '#/definitions/dto.OrderRequest'
      - description: id of the order
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновление заказа
      tags:
      - Orders
  /admins/partners:
    get:
      consumes:
      - application/json
      description: Получение партнеров
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: pagination limit
        in: query
        name: limit
        type: integer
      - description: pagination skip
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Partners'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение партнеров
      tags:
      - Partner
    post:
      consumes:
      - application/json
      description: Создание нового партнера
      parameters:
      - description: body
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/models.Partner'
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Создание нового партнера
      tags:
      - Partner
  /admins/partners/{id}:
    get:
      consumes:
      - application/json
      description: Получение партнера
      parameters:
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      - description: id of the market
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Partner'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Получение партнера
      tags:
      - Partner
    put:
      consumes:
      - application/json
      description: Обновление партнера
      parameters:
      - description: body
        in: body
        name: partner
        required: true
        schema:
          $ref: '#/definitions/dto.PartnerRequest'
      - description: id of the market
        in: path
        name: id
        required: true
        type: string
      - description: Authorization
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httperrors.Response'
        "429":
          description: Too Many Requests
          schema:
            $ref: '#/definitions/httperrors.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httperrors.Response'
      security:
      - ApiKeyAuth: []
      summary: Обновление партнера
      tags:
      - Partner
swagger: "2.0"